+++ src/titiler/core/titiler/core/models/mapbox.py
@@ -1,8 +1,16 @@
 """Common response models."""
 
+from enum import Enum
 from typing import List, Optional, Tuple
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, root_validator
+
+
+class SchemeEnum(str, Enum):
+    """TileJSON scheme choice."""
+
+    xyz = "xyz"
+    tms = "tms"
 
 
 class TileJSON(BaseModel):
@@ -20,11 +28,28 @@ class TileJSON(BaseModel):
     attribution: Optional[str]
     template: Optional[str]
     legend: Optional[str]
-    scheme: str = "xyz"
+    scheme: SchemeEnum = SchemeEnum.xyz
     tiles: List[str]
     grids: Optional[List[str]]
     data: Optional[List[str]]
     minzoom: int = Field(0, ge=0, le=30)
     maxzoom: int = Field(30, ge=0, le=30)
     bounds: List[float] = [-180, -90, 180, 90]
-    center: Tuple[float, float, int]
+    center: Optional[Tuple[float, float, int]]
+
+    @root_validator
+    def compute_center(cls, values):
+        """Compute center if it does not exist."""
+        bounds = values["bounds"]
+        if not values.get("center"):
+            values["center"] = (
+                (bounds[0] + bounds[2]) / 2,
+                (bounds[1] + bounds[3]) / 2,
+                values["minzoom"],
+            )
+        return values
+
+    class Config:
+        """TileJSON model configuration."""
+
+        use_enum_values = True